Creation of Move() Function
After several attempts, I made a move function for the user. It accepts a letter from a user and checks to see if it is in the chosen word. If it is, it replaces it. As of right now, it is only accepting lowercase letters and only allows the user to go once. It will eventually loop once I make the play function.