Number Input Error and Spaces
If the user inputs a number, it won't take off a try off of its total tries. I also changed it to there being 10 tries in total. I also added spaces to make it more clearer for the user to see what is going on when the code is being run.